refactor: move chat into vertical feature
