+++ client/src/features/chat/components/chat-workout-draft-card.tsx
