+++ client/src/features/chat/pages/chat-billing-page.test.tsx
@@ -16,7 +16,7 @@ import {
   mockRefetchFeatureAccess,
   mockSearch,
   resetChatRouteMocks,
-} from "./-chat-route-test-utils";
+} from "../test/chat-page-test-utils";
 
 describe("ChatRouteComponent", () => {
   beforeEach(resetChatRouteMocks);
