+++ client/src/features/chat/pages/chat-page.test.tsx
@@ -14,7 +14,7 @@ import {
   mockShowErrorToast,
   mockStreamMessage,
   resetChatRouteMocks,
-} from "./-chat-route-test-utils";
+} from "../test/chat-page-test-utils";
 
 describe("ChatRouteComponent", () => {
   beforeEach(resetChatRouteMocks);
