+++ client/src/features/chat/pages/chat-reconnect-page.test.tsx
@@ -10,7 +10,7 @@ import {
   mockResumeStream,
   mockShowErrorToast,
   resetChatRouteMocks,
-} from "./-chat-route-test-utils";
+} from "../test/chat-page-test-utils";
 
 describe("ChatRouteComponent", () => {
   beforeEach(resetChatRouteMocks);
