+++ client/src/features/chat/utils/chat-session-types.ts
